+++ core/narc_engine/include/buffers/Buffer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <vulkan/vulkan.h>
-#include "devices/DeviceHandler.h"
+#include "core/DeviceHandler.h"
 
 namespace narc_engine
 {
