+++ core/narc_engine/include/core/Window.h
@@ -3,7 +3,7 @@
 #define GLFW_INCLUDE_VULKAN
 #include <GLFW/glfw3.h>
 
-#include "SwapChainSupportDetails.h"
+#include "../renderer/SwapChainSupportDetails.h"
 
 namespace narc_engine
 {
