+++ core/narc_engine/include/pch.h
@@ -7,6 +7,7 @@
 #include <map>
 #include <set>
 #include <string>
+#include <string.h>
 #include <vector>
 #include <cstdint>
 #include <fstream>
