+++ core/narc_engine/src/Engine.cpp
@@ -5,70 +5,9 @@
 
 #include "data/Vertex.h"
 #include "buffers/StaggingBuffer.h"
-#include "window/Window.h"
+#include "core/Window.h"
 
 namespace narc_engine {
-    const uint32_t g_maxFramesInFlight = 2;
-
-    const std::vector<Vertex> g_vertices = {
-        {{-0.5f, -0.5f}, {1.0f, 0.0f, 0.0f}, {1.0f, 0.0f}},
-        {{0.5f, -0.5f}, {0.0f, 1.0f, 0.0f}, {0.0f, 0.0f}},
-        {{0.5f, 0.5f}, {0.0f, 0.0f, 1.0f}, {0.0f, 1.0f}},
-        {{-0.5f, 0.5f}, {1.0f, 1.0f, 1.0f}, {1.0f, 1.0f}}
-    };
-
-    const std::vector<uint16_t> g_indices = {
-        0, 1, 2, 2, 3, 0
-    };
-
-    void Engine::recordCommandBuffer(CommandBuffer* commandBuffer, uint32_t imageIndex)
-    {
-        VkCommandBufferBeginInfo beginInfo{};
-        beginInfo.sType = VK_STRUCTURE_TYPE_COMMAND_BUFFER_BEGIN_INFO;
-        beginInfo.flags = 0; // Optional
-        beginInfo.pInheritanceInfo = nullptr; // Optional
-
-        commandBuffer->begin(beginInfo);
-
-        VkRenderPassBeginInfo renderPassInfo = m_swapChain.getRenderPassBeginInfos(imageIndex);
-        VkExtent2D swapChainExtent = m_swapChain.getSwapChainExtent();
-
-        VkClearValue clearColor = {{{0.0f, 0.0f, 0.0f, 1.0f}}};
-        renderPassInfo.clearValueCount = 1;
-        renderPassInfo.pClearValues = &clearColor;
-
-        commandBuffer->cmdBeginRenderPass(&renderPassInfo, VK_SUBPASS_CONTENTS_INLINE);
-        commandBuffer->cmdBindPipeline(VK_PIPELINE_BIND_POINT_GRAPHICS, &m_graphicsPipeline);
-
-        VkViewport viewport{};
-        viewport.x = 0.0f;
-        viewport.y = 0.0f;
-        viewport.width = (float) swapChainExtent.width;
-        viewport.height = (float) swapChainExtent.height;
-        viewport.minDepth = 0.0f;
-        viewport.maxDepth = 1.0f;
-        commandBuffer->cmdSetViewport(&viewport, 0, 1);
-
-        VkRect2D scissor{};
-        scissor.offset = {0, 0};
-        scissor.extent = swapChainExtent;
-        commandBuffer->cmdSetScissor(&scissor, 0, 1);
-
-        VkBuffer vertexBuffers[] = {m_vertexBuffer.getBuffer()};
-        VkDeviceSize offsets[] = {0};
-        commandBuffer->cmdBindVertexBuffers(0, 1, vertexBuffers, offsets);
-        commandBuffer->cmdBindIndexBuffer(m_indexBuffer.getBuffer(), 0, VK_INDEX_TYPE_UINT16);
-        commandBuffer->cmdBindDescriptorSets(VK_PIPELINE_BIND_POINT_GRAPHICS, &m_graphicsPipeline, 0, 1, &m_descriptorSets[m_currentFrame], 0, nullptr);
-
-        commandBuffer->cmdDrawIndexed(static_cast<uint32_t>(g_indices.size()), 1, 0, 0, 0);
-        commandBuffer->cmdEndRenderPass();
-
-        if (commandBuffer->end() != VK_SUCCESS)
-        {
-            throw std::runtime_error("Failed to record command buffer!");
-        }
-    }
-
     static Engine* s_instance;
 
     Engine::Engine()
@@ -79,55 +18,16 @@ namespace narc_engine {
         m_debugLogger.init(m_vulkanInstance);
         m_window.initSurface(m_vulkanInstance);
         m_deviceHandler.create(&m_window, m_vulkanInstance, m_debugLogger);
-        m_swapChain.create(); // CreateSwapChain(); // CreateImageViews(); //CreateRenderPass();
-        createDescriptorSetLayout();
-        m_graphicsPipeline.create(&m_swapChain, &m_descriptorSetLayout);
-        m_swapChain.createFramebuffers(); // CreateFramebuffers();
         m_commandPool.create(&m_deviceHandler);
-        createTextureImage();
-        createImageTextureView();
-        createTextureSampler();
-        m_vertexBuffer.create(g_vertices, VK_BUFFER_USAGE_VERTEX_BUFFER_BIT);
-        m_indexBuffer.create(g_indices, VK_BUFFER_USAGE_INDEX_BUFFER_BIT);
-        createUniformBuffers();
-        createDescriptorPool();
-        createDescriptorSets();
-        m_commandPool.createCommandBuffers(g_maxFramesInFlight);
-        createSyncObjects();
+        m_renderer.create();
     }
 
     Engine::~Engine()
     {
-        const VkDevice& device = m_deviceHandler.getDevice();
+        m_renderer.release();
+        // m_swapChain.cleanSwapChain();
 
-        m_swapChain.cleanSwapChain();
 
-        vkDestroySampler(device, m_textureSampler, nullptr);
-        vkDestroyImageView(device, m_textureImageView, nullptr);
-
-        vkDestroyImage(device, m_textureImage, nullptr);
-        vkFreeMemory(device, m_textureImageMemory, nullptr);
-
-        for (UniformBuffer& buffer: m_uniformBuffers)
-        {
-            buffer.release();
-        }
-
-        m_descriptorPool.release();
-        vkDestroyDescriptorSetLayout(device, m_descriptorSetLayout, nullptr);
-
-        m_indexBuffer.release();
-        m_vertexBuffer.release();
-
-        m_graphicsPipeline.release();
-        m_swapChain.cleanRenderPass();
-
-        for (size_t i = 0; i < g_maxFramesInFlight; i++)
-        {
-            vkDestroySemaphore(device, m_renderFinishedSemaphores[i], nullptr);
-            vkDestroySemaphore(device, m_imageAvailableSemaphores[i], nullptr);
-            vkDestroyFence(device, m_inFlightFences[i], nullptr);
-        }
 
         m_commandPool.release();
         m_deviceHandler.release();
@@ -142,17 +42,6 @@ namespace narc_engine {
     Engine* Engine::getInstance()
     {
         return s_instance;
-    }
-
-    void Engine::mainLoop()
-    {
-        while (!m_window.shouldClose())
-        {
-            m_window.update();
-            drawFrame();
-        }
-
-        //m_deviceHandler.waitDeviceIdle();
     }
 
     void Engine::createVulkanInstance()
@@ -187,197 +76,6 @@ namespace narc_engine {
         {
             throw std::runtime_error("Failed to create instance!");
         }
-    }
-
-    void Engine::createUniformBuffers()
-    {
-        VkDeviceSize bufferSize = sizeof(UniformBufferObject);
-
-        m_uniformBuffers.resize(g_maxFramesInFlight);
-
-        for (size_t i = 0; i < g_maxFramesInFlight; i++)
-        {
-            m_uniformBuffers[i].create(bufferSize);
-        }
-    }
-
-    void Engine::createDescriptorPool()
-    {
-        uint32_t descriptionCount = g_maxFramesInFlight;
-        DescriptorPoolBuilder builder;
-        builder.addPoolSize(VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER, descriptionCount);
-        builder.addPoolSize(VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER, descriptionCount);
-        builder.setMaxSet(descriptionCount);
-
-        m_descriptorPool.create(&builder);
-    }
-
-    void Engine::createDescriptorSets()
-    {
-        std::vector<VkDescriptorSetLayout> layouts(g_maxFramesInFlight, m_descriptorSetLayout);
-        VkDescriptorSetAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_ALLOCATE_INFO;
-        allocInfo.descriptorSetCount = g_maxFramesInFlight;
-        allocInfo.pSetLayouts = layouts.data();
-
-        m_descriptorSets.resize(g_maxFramesInFlight);
-        m_descriptorPool.allocateDescriptorSets(&allocInfo, m_descriptorSets.data());
-
-        for (size_t i = 0; i < g_maxFramesInFlight; i++)
-        {
-            VkDescriptorBufferInfo bufferInfo{};
-            bufferInfo.buffer = m_uniformBuffers[i].getBuffer();
-            bufferInfo.offset = 0;
-            bufferInfo.range = sizeof(UniformBufferObject);
-
-            VkDescriptorImageInfo imageInfo{};
-            imageInfo.imageLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-            imageInfo.imageView = m_textureImageView;
-            imageInfo.sampler = m_textureSampler;
-
-            std::array<VkWriteDescriptorSet, 2> descriptorWrites{};
-
-            descriptorWrites[0].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            descriptorWrites[0].dstSet = m_descriptorSets[i];
-            descriptorWrites[0].dstBinding = 0;
-            descriptorWrites[0].dstArrayElement = 0;
-            descriptorWrites[0].descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-            descriptorWrites[0].descriptorCount = 1;
-            descriptorWrites[0].pBufferInfo = &bufferInfo;
-
-            descriptorWrites[1].sType = VK_STRUCTURE_TYPE_WRITE_DESCRIPTOR_SET;
-            descriptorWrites[1].dstSet = m_descriptorSets[i];
-            descriptorWrites[1].dstBinding = 1;
-            descriptorWrites[1].dstArrayElement = 0;
-            descriptorWrites[1].descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-            descriptorWrites[1].descriptorCount = 1;
-            descriptorWrites[1].pImageInfo = &imageInfo;
-
-            vkUpdateDescriptorSets(m_deviceHandler.getDevice(), static_cast<uint32_t>(descriptorWrites.size()), descriptorWrites.data(), 0, nullptr);
-        }
-    }
-
-    void Engine::createDescriptorSetLayout()
-    {
-        VkDescriptorSetLayoutBinding uboLayoutBinding{};
-        uboLayoutBinding.binding = 0;
-        uboLayoutBinding.descriptorType = VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER;
-        uboLayoutBinding.descriptorCount = 1;
-        uboLayoutBinding.stageFlags = VK_SHADER_STAGE_VERTEX_BIT;
-        uboLayoutBinding.pImmutableSamplers = nullptr;
-
-        VkDescriptorSetLayoutBinding samplerLayoutBinding{};
-        samplerLayoutBinding.binding = 1;
-        samplerLayoutBinding.descriptorCount = 1;
-        samplerLayoutBinding.descriptorType = VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER;
-        samplerLayoutBinding.pImmutableSamplers = nullptr;
-        samplerLayoutBinding.stageFlags = VK_SHADER_STAGE_FRAGMENT_BIT;
-
-        std::array<VkDescriptorSetLayoutBinding, 2> bindings = {uboLayoutBinding, samplerLayoutBinding};
-        VkDescriptorSetLayoutCreateInfo layoutInfo{};
-        layoutInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
-        layoutInfo.bindingCount = static_cast<uint32_t>(bindings.size());
-        layoutInfo.pBindings = bindings.data();
-
-        if (vkCreateDescriptorSetLayout(m_deviceHandler.getDevice(), &layoutInfo, nullptr, &m_descriptorSetLayout) != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to create descriptor set layout!");
-        }
-    }
-
-    void Engine::createTextureImage()
-    {
-        narc_io::Image image = narc_io::FileReader::readImage("textures/logo.png");
-        VkDeviceSize imageSize = image.getWidth() * image.getHeight() * 4;
-
-        StaggingBuffer staggingBuffer;
-        staggingBuffer.create(imageSize);
-        staggingBuffer.input(image.getData());
-
-        // stbi_image_free(pixels);
-
-        createImage(image, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_TILING_OPTIMAL, VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT, VK_MEMORY_PROPERTY_DEVICE_LOCAL_BIT,
-                    m_textureImage, m_textureImageMemory);
-
-        transitionImageLayout(m_textureImage, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_LAYOUT_UNDEFINED, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL);
-        copyBufferToImage(staggingBuffer.getBuffer(), m_textureImage, static_cast<uint32_t>(image.getWidth()), static_cast<uint32_t>(image.getHeight()));
-        transitionImageLayout(m_textureImage, VK_FORMAT_R8G8B8A8_SRGB, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL);
-
-        staggingBuffer.release();
-    }
-
-    void Engine::createTextureSampler()
-    {
-        VkSamplerCreateInfo samplerInfo{};
-        samplerInfo.sType = VK_STRUCTURE_TYPE_SAMPLER_CREATE_INFO;
-        samplerInfo.magFilter = VK_FILTER_LINEAR;
-        samplerInfo.minFilter = VK_FILTER_LINEAR;
-        samplerInfo.addressModeU = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-        samplerInfo.addressModeV = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-        samplerInfo.addressModeW = VK_SAMPLER_ADDRESS_MODE_REPEAT;
-
-        VkPhysicalDeviceProperties properties = m_deviceHandler.getPhysicalDeviceProperties();
-        samplerInfo.anisotropyEnable = VK_TRUE;
-        samplerInfo.maxAnisotropy = properties.limits.maxSamplerAnisotropy;
-
-        samplerInfo.borderColor = VK_BORDER_COLOR_INT_OPAQUE_BLACK;
-        samplerInfo.unnormalizedCoordinates = VK_FALSE;
-        samplerInfo.compareEnable = VK_FALSE;
-        samplerInfo.compareOp = VK_COMPARE_OP_ALWAYS;
-        samplerInfo.mipmapMode = VK_SAMPLER_MIPMAP_MODE_LINEAR;
-        samplerInfo.mipLodBias = 0.0f;
-        samplerInfo.minLod = 0.0f;
-        samplerInfo.maxLod = 0.0f;
-
-        if (vkCreateSampler(m_deviceHandler.getDevice(), &samplerInfo, nullptr, &m_textureSampler) != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to create texture sampler!");
-        }
-    }
-
-    void Engine::createImageTextureView()
-    {
-        m_textureImageView = m_swapChain.createImageView(m_textureImage, VK_FORMAT_R8G8B8A8_SRGB);
-    }
-
-    void Engine::createImage(const narc_io::Image& imageData, VkFormat format, VkImageTiling tiling, VkImageUsageFlags usage, VkMemoryPropertyFlags properties, VkImage& image,
-                             VkDeviceMemory& imageMemory)
-    {
-        VkImageCreateInfo imageInfo{};
-        imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
-        imageInfo.imageType = VK_IMAGE_TYPE_2D;
-        imageInfo.extent.width = imageData.getWidth();
-        imageInfo.extent.height = imageData.getHeight();
-        imageInfo.extent.depth = 1;
-        imageInfo.mipLevels = 1;
-        imageInfo.arrayLayers = 1;
-        imageInfo.format = format;
-        imageInfo.tiling = tiling;
-        imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-        imageInfo.usage = usage;
-        imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
-        imageInfo.samples = VK_SAMPLE_COUNT_1_BIT;
-        imageInfo.flags = 0; // Optional
-
-        if (vkCreateImage(m_deviceHandler.getDevice(), &imageInfo, nullptr, &image) != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to create image!");
-        }
-
-        VkMemoryRequirements memRequirements;
-        vkGetImageMemoryRequirements(m_deviceHandler.getDevice(), image, &memRequirements);
-
-        VkMemoryAllocateInfo allocInfo{};
-        allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
-        allocInfo.allocationSize = memRequirements.size;
-        allocInfo.memoryTypeIndex = m_deviceHandler.findMemoryType(memRequirements.memoryTypeBits, properties);
-
-        if (vkAllocateMemory(m_deviceHandler.getDevice(), &allocInfo, nullptr, &imageMemory) != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to allocate image memory!");
-        }
-
-        vkBindImageMemory(m_deviceHandler.getDevice(), image, imageMemory, 0);
     }
 
     void Engine::transitionImageLayout(VkImage image, VkFormat format, VkImageLayout oldLayout, VkImageLayout newLayout)
@@ -464,110 +162,43 @@ namespace narc_engine {
         m_commandPool.endSingleTimeCommands(commandBuffer);
     }
 
-    void Engine::createSyncObjects()
+    void Engine::createImage(const narc_io::Image& imageData, VkFormat format, VkImageTiling tiling, VkImageUsageFlags usage, VkMemoryPropertyFlags properties, VkImage& image,
+                             VkDeviceMemory& imageMemory) const
     {
-        m_imageAvailableSemaphores.resize(g_maxFramesInFlight);
-        m_renderFinishedSemaphores.resize(g_maxFramesInFlight);
-        m_inFlightFences.resize(g_maxFramesInFlight);
+        VkImageCreateInfo imageInfo{};
+        imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
+        imageInfo.imageType = VK_IMAGE_TYPE_2D;
+        imageInfo.extent.width = imageData.getWidth();
+        imageInfo.extent.height = imageData.getHeight();
+        imageInfo.extent.depth = 1;
+        imageInfo.mipLevels = 1;
+        imageInfo.arrayLayers = 1;
+        imageInfo.format = format;
+        imageInfo.tiling = tiling;
+        imageInfo.initialLayout = VK_IMAGE_LAYOUT_UNDEFINED;
+        imageInfo.usage = usage;
+        imageInfo.sharingMode = VK_SHARING_MODE_EXCLUSIVE;
+        imageInfo.samples = VK_SAMPLE_COUNT_1_BIT;
+        imageInfo.flags = 0; // Optional
 
-        VkSemaphoreCreateInfo semaphoreInfo{};
-        semaphoreInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-
-        VkFenceCreateInfo fenceInfo{};
-        fenceInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-        fenceInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-
-        for (size_t i = 0; i < g_maxFramesInFlight; i++)
+        if (vkCreateImage(m_deviceHandler.getDevice(), &imageInfo, nullptr, &image) != VK_SUCCESS)
         {
-            if (vkCreateSemaphore(m_deviceHandler.getDevice(), &semaphoreInfo, nullptr, &m_imageAvailableSemaphores[i]) != VK_SUCCESS ||
-                vkCreateSemaphore(m_deviceHandler.getDevice(), &semaphoreInfo, nullptr, &m_renderFinishedSemaphores[i]) != VK_SUCCESS ||
-                vkCreateFence(m_deviceHandler.getDevice(), &fenceInfo, nullptr, &m_inFlightFences[i]) != VK_SUCCESS)
-            {
-                throw std::runtime_error("failed to create semaphores!");
-            }
-        }
-    }
-
-    void Engine::drawFrame()
-    {
-        vkWaitForFences(m_deviceHandler.getDevice(), 1, &m_inFlightFences[m_currentFrame], VK_TRUE, UINT64_MAX);
-
-        uint32_t imageIndex;
-        VkResult result = m_swapChain.acquireNextImage(m_imageAvailableSemaphores[m_currentFrame], &imageIndex);
-
-        updateUniformBuffer(m_currentFrame);
-
-        vkResetFences(m_deviceHandler.getDevice(), 1, &m_inFlightFences[m_currentFrame]);
-
-        CommandBuffer* buffer = m_commandPool.getCommandBuffer(m_currentFrame);
-        buffer->reset(0);
-        recordCommandBuffer(buffer, imageIndex);
-
-        VkSubmitInfo submitInfo{};
-        submitInfo.sType = VK_STRUCTURE_TYPE_SUBMIT_INFO;
-
-        VkSemaphore waitSemaphores[] =
-        {
-            m_imageAvailableSemaphores[m_currentFrame]
-        };
-        VkPipelineStageFlags waitStages[] =
-        {
-            VK_PIPELINE_STAGE_COLOR_ATTACHMENT_OUTPUT_BIT
-        };
-        submitInfo.waitSemaphoreCount = 1;
-        submitInfo.pWaitSemaphores = waitSemaphores;
-        submitInfo.pWaitDstStageMask = waitStages;
-        submitInfo.commandBufferCount = 1;
-        submitInfo.pCommandBuffers = buffer->getVkCommandBuffer();
-
-        VkSemaphore signalSemaphores[] = {m_renderFinishedSemaphores[m_currentFrame]};
-        submitInfo.signalSemaphoreCount = 1;
-        submitInfo.pSignalSemaphores = signalSemaphores;
-
-        if (m_deviceHandler.submitGraphicsQueue(1, &submitInfo, m_inFlightFences[m_currentFrame]) != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to submit draw command buffer!");
+            throw std::runtime_error("failed to create image!");
         }
 
-        VkPresentInfoKHR presentInfo{};
-        presentInfo.sType = VK_STRUCTURE_TYPE_PRESENT_INFO_KHR;
-        presentInfo.waitSemaphoreCount = 1;
-        presentInfo.pWaitSemaphores = signalSemaphores;
+        VkMemoryRequirements memRequirements;
+        vkGetImageMemoryRequirements(m_deviceHandler.getDevice(), image, &memRequirements);
 
-        VkSwapchainKHR swapChains[] = {m_swapChain.getSwapChain()};
-        presentInfo.swapchainCount = 1;
-        presentInfo.pSwapchains = swapChains;
+        VkMemoryAllocateInfo allocInfo{};
+        allocInfo.sType = VK_STRUCTURE_TYPE_MEMORY_ALLOCATE_INFO;
+        allocInfo.allocationSize = memRequirements.size;
+        allocInfo.memoryTypeIndex = m_deviceHandler.findMemoryType(memRequirements.memoryTypeBits, properties);
 
-        presentInfo.pImageIndices = &imageIndex;
-        presentInfo.pResults = nullptr;
-
-        result = m_deviceHandler.presentKHR(&presentInfo);
-
-        if (result == VK_ERROR_OUT_OF_DATE_KHR || result == VK_SUBOPTIMAL_KHR || m_window.isFramebufferResized())
+        if (vkAllocateMemory(m_deviceHandler.getDevice(), &allocInfo, nullptr, &imageMemory) != VK_SUCCESS)
         {
-            m_window.setFramebufferResized(false); //after vkQueuePresentKHR to ensure that the semaphores are in a consistent state
-            m_swapChain.reCreate();
-        } else if (result != VK_SUCCESS)
-        {
-            throw std::runtime_error("failed to present swap chain image!");
+            throw std::runtime_error("failed to allocate image memory!");
         }
 
-        m_currentFrame = (m_currentFrame + 1) % g_maxFramesInFlight;
-    }
-
-    void Engine::updateUniformBuffer(uint32_t currentImage)
-    {
-        static auto startTime = std::chrono::high_resolution_clock::now();
-
-        auto currentTime = std::chrono::high_resolution_clock::now();
-        float time = std::chrono::duration<float, std::chrono::seconds::period>(currentTime - startTime).count();
-
-        UniformBufferObject ubo{};
-        ubo.Model = glm::rotate(glm::mat4(1.0f), time * glm::radians(90.0f), glm::vec3(0.0f, 0.0f, 1.0f));
-        ubo.View = glm::lookAt(glm::vec3(2.0f, 2.0f, 2.0f), glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3(0.0f, 0.0f, 1.0f));
-        ubo.Proj = glm::perspective(glm::radians(45.0f), m_swapChain.getSwapChainExtent().width / (float) m_swapChain.getSwapChainExtent().height, 0.1f, 10.0f);
-        ubo.Proj[1][1] *= -1;
-
-        m_uniformBuffers[currentImage].setData(ubo);
+        vkBindImageMemory(m_deviceHandler.getDevice(), image, imageMemory, 0);
     }
 }
