+++ core/narc_engine/src/Pipeline.cpp
@@ -1,4 +1,4 @@
-#include "Pipeline.h"
+#include "../include/renderer/Pipeline.h"
 
 #include "Engine.h"
 #include "data/Vertex.h"
