+++ core/narc_engine/src/renderer/DescriptorPool.cpp
@@ -1,6 +1,5 @@
-#include "DescriptorPool.h"
+#include "renderer/DescriptorPool.h"
 
-#include "DescriptorPoolBuilder.h"
 #include "Engine.h"
 
 namespace narc_engine {
