+++ core/narc_engine/src/renderer/SwapChain.cpp
@@ -1,7 +1,7 @@
-#include "SwapChain.h"
+#include "renderer/SwapChain.h"
 
 #include "Engine.h"
-#include "devices/DeviceHandler.h"
+#include "core/DeviceHandler.h"
 
 namespace narc_engine
 {
