+++ core/narcio/include/FileReader.h
@@ -1,5 +1,7 @@
 ﻿#pragma once
 
+#include "Core.h"
+
 namespace narc_io
 {
     struct Image;
